# Faster protobuf parsing