## Comparison of running times 

All benchmarks were run using `go test ./parser -bench . -benchtime 10s` on a Core i9-9980HK.

`master`

| Parser | ops | ns/op |
|:--- | ---:| ---:| 
| BenchmarkJSONParse-16 | 3138224 | 3829 ns/op |
| BenchmarkProtoBufParse-16 | 307641 | 39706 ns/op |

Initial optimization (extract message only and parse that using jsonpb)

| Parser | ops | ns/op |
|:--- | ---:| ---:| 
| BenchmarkJSONParse-16 | 2980334 | 3965 ns/op |
| BenchmarkProtoBufParse-16 | 533475 | 18993 ns/op |

Use encoding/json parser (this change)

| Parser | ops | ns/op |
|:--- | ---:| ---:| 
| BenchmarkJSONParse-16 | 2913828 | 3886 ns/op |
| BenchmarkProtoBufParse-16 | 1243095 | 9672 ns/op |